refactor: 등락률 타입 변환 (int->float)

+++ backend/src/main/java/com/example/backend/dto/DailyPriceStockNameDTO.java
@@ -24,7 +24,7 @@ public class DailyPriceStockNameDTO {
 
     private Integer open;  // 시가
 
-    private Integer changeRate; // 등락율
+    private Float changeRate; // 등락율
 
     private Integer volume; // 일거래량
 }
